+++ backend/src/test/java/org/donggle/backend/application/service/PublishFacadeServiceTest.java
@@ -1,8 +1,8 @@
 package org.donggle.backend.application.service;
 
-import org.donggle.backend.application.service.blog.PublishFacadeService;
-import org.donggle.backend.application.service.blog.PublishService;
-import org.donggle.backend.application.service.blog.PublishWritingRequest;
+import org.donggle.backend.application.service.publish.PublishFacadeService;
+import org.donggle.backend.application.service.publish.PublishService;
+import org.donggle.backend.application.service.publish.PublishWritingRequest;
 import org.donggle.backend.application.service.request.PublishRequest;
 import org.donggle.backend.domain.blog.Blog;
 import org.donggle.backend.domain.blog.BlogClients;
