+++ backend/src/test/java/org/donggle/backend/ui/ControllerTest.java
@@ -4,9 +4,9 @@ package org.donggle.backend.ui;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.donggle.backend.application.repository.TokenRepository;
 import org.donggle.backend.application.service.auth.AuthFacadeService;
-import org.donggle.backend.application.service.blog.PublishFacadeService;
 import org.donggle.backend.application.service.category.CategoryService;
 import org.donggle.backend.application.service.member.MemberService;
+import org.donggle.backend.application.service.publish.PublishFacadeService;
 import org.donggle.backend.application.service.trash.TrashService;
 import org.donggle.backend.application.service.writing.WritingFacadeService;
 import org.donggle.backend.domain.auth.JwtTokenProvider;
